refactor: extract chat session utilities
